Add files via upload
Task-1:
Today we explored following two sensors:
● Temperature sensor
● Soil moisture sensor
Using these two sensors you need to implement a system that does the following:
1. Print the temperature value in LCD (-40
oC
to 125
oC)
2. Print moisture value in percentage (0% to 100%)
3. Play a buzzer as an alarm when both sensors’ values go below 30 (⩽30). Meaning when
both the temperature is below 30

oC and humidity is below 30% the buzzer will ring. If

individually any of them do below 30, the buzzer will not ring.